Remove unused code from base-android
There was very little left, so removed the module
and created 2 provider modules for analytics + logging.
