+++ app/src/main/java/app/tivi/appinitializers/AppInitializers.kt
@@ -16,15 +16,14 @@
 
 package app.tivi.appinitializers
 
-import android.app.Application
 import javax.inject.Inject
 
 class AppInitializers @Inject constructor(
     private val initializers: Set<@JvmSuppressWildcards AppInitializer>
 ) {
-    fun init(application: Application) {
+    fun init() {
         initializers.forEach {
-            it.init(application)
+            it.init()
         }
     }
 }
