+++ app/src/main/java/app/tivi/appinitializers/TmdbInitializer.kt
@@ -16,7 +16,6 @@
 
 package app.tivi.appinitializers
 
-import android.app.Application
 import app.tivi.domain.interactors.UpdateTmdbConfig
 import app.tivi.util.AppCoroutineDispatchers
 import kotlinx.coroutines.DelicateCoroutinesApi
@@ -28,7 +27,7 @@ class TmdbInitializer @Inject constructor(
     private val updateTmdbConfig: UpdateTmdbConfig,
     private val dispatchers: AppCoroutineDispatchers
 ) : AppInitializer {
-    override fun init(application: Application) {
+    override fun init() {
         @OptIn(DelicateCoroutinesApi::class)
         GlobalScope.launch(dispatchers.main) {
             updateTmdbConfig.executeSync(Unit)
