--- base-android/sampledata/episode_summary.txt
@@ -1,1 +1,0 @@
-An episode about a street cat.
