--- base-android/sampledata/shows.txt
@@ -1,20 +1,0 @@
-Game of Thrones
-Breaking Bad
-The Walking Dead
-The Big Bang Theory
-Sherlock
-Dexter
-How I Met Your Mother
-Arrow
-Friends
-Lost
-House
-Homeland
-House of Cards
-Fringe
-Suits
-Supernatural
-Modern Family
-Stranger Things
-Orange Is the New Black
-Doctor Who
