--- base-android/src/main/java/app/tivi/BaseModule.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2021 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package app.tivi
-
-import app.tivi.util.Analytics
-import app.tivi.util.Logger
-import app.tivi.util.TiviAnalytics
-import app.tivi.util.TiviLogger
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-abstract class BaseModule {
-    @Singleton
-    @Binds
-    internal abstract fun provideLogger(bind: TiviLogger): Logger
-
-    @Singleton
-    @Binds
-    internal abstract fun provideAnalytics(bind: TiviAnalytics): Analytics
-}
