--- base-android/src/main/java/app/tivi/extensions/ContextExtensions.kt
@@ -1,49 +1,0 @@
-/*
- * Copyright 2019 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package app.tivi.extensions
-
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
-import android.view.inputmethod.InputMethodManager
-import androidx.core.content.getSystemService
-import androidx.fragment.app.Fragment
-
-inline fun <reified T : Context> Context.findBaseContext(): T? {
-    var ctx: Context? = this
-    do {
-        if (ctx is T) {
-            return ctx
-        }
-        if (ctx is ContextWrapper) {
-            ctx = ctx.baseContext
-        }
-    } while (ctx != null)
-
-    // If we reach here, there's not an Context of type T in our Context hierarchy
-    return null
-}
-
-fun Activity.hideSoftInput() {
-    val imm: InputMethodManager? = getSystemService()
-    val currentFocus = currentFocus
-    if (currentFocus != null && imm != null) {
-        imm.hideSoftInputFromWindow(currentFocus.windowToken, 0)
-    }
-}
-
-fun Fragment.hideSoftInput() = requireActivity().hideSoftInput()
