+++ common-imageloading/build.gradle
@@ -40,7 +40,6 @@ android {
 
 dependencies {
     implementation project(':base')
-    implementation project(':base-android')
     implementation project(':data')
     implementation project(':common-ui-view')
     implementation project(':tmdb')
