+++ core-analytics/build.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,7 +30,6 @@ android {
 
     defaultConfig {
         minSdkVersion buildConfig.minSdk
-        manifestPlaceholders += ['appAuthRedirectScheme': 'empty']
     }
 
     compileOptions {
@@ -40,15 +39,10 @@ android {
 }
 
 dependencies {
-    implementation project(':base')
-    implementation project(':data')
-    api project(':trakt')
+    api project(':base')
+
+    implementation libs.google.analytics
 
     implementation libs.hilt.library
     kapt libs.hilt.compiler
-
-    implementation libs.androidx.browser
-    implementation libs.androidx.core
-
-    api libs.appauth
 }
