+++ domain/build.gradle
@@ -40,7 +40,6 @@ android {
 
 dependencies {
     implementation project(':base')
-    implementation project(':base-android')
     api project(':data')
     implementation project(':data-android')
     implementation project(':trakt-auth')
