+++ ui-account/build.gradle
@@ -49,7 +49,6 @@ android {
 
 dependencies {
     implementation project(':base')
-    implementation project(':base-android')
     implementation project(':domain')
     implementation project(':common-ui-view')
     implementation project(':common-ui-compose')
