+++ ui-episodedetails/build.gradle
@@ -50,7 +50,6 @@ android {
 
 dependencies {
     implementation project(':base')
-    implementation project(':base-android')
     implementation project(':domain')
     implementation project(':common-ui-view')
     implementation project(':common-ui-compose')
