+++ ui-settings/build.gradle
@@ -42,7 +42,6 @@ android {
 
 dependencies {
     implementation project(':base')
-    implementation project(':base-android')
     implementation project(':common-ui-resources')
     implementation project(':common-ui-view')
